Move CLI into package
